--- domain/src/main/java/com/tago/domain/auth/domain/OAuthProvider.java
@@ -1,6 +1,0 @@
-package com.tago.domain.auth.domain;
-
-public enum OAuthProvider {
-
-    KAKAO
-}
